+++ src/main/java/repositories/interfaces/NewsRepository.java
@@ -12,7 +12,7 @@ public interface NewsRepository extends EntityRepository<News> {
 
     List<News> getNewsByUser(User user);
 
-    News getNewsById(int id);
+    News getNewsById(long id);
 
     User getNewsAuthor(News news);
